Rename files in model folder
