+++ docs/interfaces/Desk.html
@@ -107,7 +107,7 @@
 					<div class="tsd-signature tsd-kind-icon">id<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Desk.ts#L19">model/Desk.ts:19</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Desk.ts#L19">model/desk.ts:19</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -123,7 +123,7 @@
 					<div class="tsd-signature tsd-kind-icon">name<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Desk.ts#L26">model/Desk.ts:26</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Desk.ts#L26">model/desk.ts:26</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
