+++ docs/interfaces/Device.html
@@ -107,7 +107,7 @@
 					<div class="tsd-signature tsd-kind-icon">battery<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol">{ </span>charging<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">boolean</span><span class="tsd-signature-symbol">; </span>level<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">number</span><span class="tsd-signature-symbol"> }</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Device.ts#L49">model/Device.ts:49</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Device.ts#L49">model/device.ts:49</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -141,7 +141,7 @@
 					<div class="tsd-signature tsd-kind-icon">id<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Device.ts#L7">model/Device.ts:7</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Device.ts#L7">model/device.ts:7</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -156,7 +156,7 @@
 					<div class="tsd-signature tsd-kind-icon">name<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Device.ts#L26">model/Device.ts:26</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Device.ts#L26">model/device.ts:26</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -171,7 +171,7 @@
 					<div class="tsd-signature tsd-kind-icon">needs<wbr>Update<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">boolean</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Device.ts#L32">model/Device.ts:32</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Device.ts#L32">model/device.ts:32</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -186,7 +186,7 @@
 					<div class="tsd-signature tsd-kind-icon">offline<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">boolean</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Device.ts#L29">model/Device.ts:29</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Device.ts#L29">model/device.ts:29</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -201,7 +201,7 @@
 					<div class="tsd-signature tsd-kind-icon">type<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">&quot;MONITOR&quot;</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">&quot;OVERVIEW_MONITOR&quot;</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">&quot;PRINTER&quot;</span><span class="tsd-signature-symbol"> | </span><span class="tsd-signature-type">&quot;NAME_DEVICE&quot;</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Device.ts#L23">model/Device.ts:23</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Device.ts#L23">model/device.ts:23</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
