+++ docs/interfaces/Line.html
@@ -109,7 +109,7 @@
 					<div class="tsd-signature tsd-kind-icon">color<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Line.ts#L27">model/Line.ts:27</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Line.ts#L27">model/line.ts:27</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -125,7 +125,7 @@
 					<div class="tsd-signature tsd-kind-icon">disabled<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">boolean</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Line.ts#L32">model/Line.ts:32</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Line.ts#L32">model/line.ts:32</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -141,7 +141,7 @@
 					<div class="tsd-signature tsd-kind-icon">id<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Line.ts#L16">model/Line.ts:16</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Line.ts#L16">model/line.ts:16</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -157,7 +157,7 @@
 					<div class="tsd-signature tsd-kind-icon">name<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Line.ts#L22">model/Line.ts:22</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Line.ts#L22">model/line.ts:22</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
