+++ docs/interfaces/Location.html
@@ -108,7 +108,7 @@
 					<div class="tsd-signature tsd-kind-icon">country<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Location.ts#L36">model/Location.ts:36</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Location.ts#L36">model/location.ts:36</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -124,7 +124,7 @@
 					<div class="tsd-signature tsd-kind-icon">id<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Location.ts#L10">model/Location.ts:10</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Location.ts#L10">model/location.ts:10</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -139,7 +139,7 @@
 					<div class="tsd-signature tsd-kind-icon">latitude<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Location.ts#L25">model/Location.ts:25</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Location.ts#L25">model/location.ts:25</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -155,7 +155,7 @@
 					<div class="tsd-signature tsd-kind-icon">longitude<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Location.ts#L30">model/Location.ts:30</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Location.ts#L30">model/location.ts:30</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -171,7 +171,7 @@
 					<div class="tsd-signature tsd-kind-icon">name<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Location.ts#L14">model/Location.ts:14</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Location.ts#L14">model/location.ts:14</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -186,7 +186,7 @@
 					<div class="tsd-signature tsd-kind-icon">timezone<wbr>Offset<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Location.ts#L20">model/Location.ts:20</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/Location.ts#L20">model/location.ts:20</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
