+++ docs/interfaces/User.html
@@ -111,7 +111,7 @@
 					<div class="tsd-signature tsd-kind-icon">desk<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/User.ts#L63">model/User.ts:63</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/User.ts#L63">model/user.ts:63</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -126,7 +126,7 @@
 					<div class="tsd-signature tsd-kind-icon">email<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/User.ts#L49">model/User.ts:49</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/User.ts#L49">model/user.ts:49</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -143,7 +143,7 @@
 					<div class="tsd-signature tsd-kind-icon">first<wbr>Name<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/User.ts#L54">model/User.ts:54</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/User.ts#L54">model/user.ts:54</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -159,7 +159,7 @@
 					<div class="tsd-signature tsd-kind-icon">id<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/User.ts#L43">model/User.ts:43</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/User.ts#L43">model/user.ts:43</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -175,7 +175,7 @@
 					<div class="tsd-signature tsd-kind-icon">last<wbr>Name<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/User.ts#L59">model/User.ts:59</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/User.ts#L59">model/user.ts:59</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -191,7 +191,7 @@
 					<div class="tsd-signature tsd-kind-icon">picture<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">Picture</span><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/User.ts#L71">model/User.ts:71</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/User.ts#L71">model/user.ts:71</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -206,7 +206,7 @@
 					<div class="tsd-signature tsd-kind-icon">roles<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">UserRole</span><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/User.ts#L75">model/User.ts:75</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/User.ts#L75">model/user.ts:75</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -221,7 +221,7 @@
 					<div class="tsd-signature tsd-kind-icon">selected<wbr>Location<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">number</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/User.ts#L67">model/User.ts:67</a></li>
+							<li>Defined in <a href="https://github.com/Qminder/javascript-api/blob/5cc8a56/src/model/User.ts#L67">model/user.ts:67</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
