+++ src/api-base.ts
@@ -1,6 +1,6 @@
 import fetch from 'cross-fetch';
 import { GraphQLApiError } from './util/errors.js';
-import { ClientError } from './model/ClientError.js';
+import { ClientError } from './model/client-error';
 
 type HTTPMethod =
   | 'GET'
