--- src/model/ID.ts
@@ -1,1 +1,0 @@
-export type ID = string;
