+++ src/model/id.ts
@@ -1,0 +1,1 @@
+export type Id = string;
