+++ src/model/ticket.ts
