+++ src/model/user.ts
@@ -1,4 +1,3 @@
-import Location from './Location.js';
 /**
  * Represents a single user picture.
  * The 'medium' sized user picture is available if the user has an image.
